Make the time to wait for the stacktrace after a lock obain failure configurable in the MonitorCE class..
Change the default to 1 second. We are primarily trying to diagnose in memory deadlocks with this.
If that is what is going on, 1 second is plenty. If it is not an in memory deadlock, we are preventing logging
frameworks etc from getting on with their jobs by refusing to return the message property.